split UNO into modules
