+++ games/UNO.js
@@ -1,8 +1,8 @@
+import {Card} from './structures/UNO/Card';
 import {Game} from './structures/Game';
-import {AbstractPlayer} from './structures/AbstractPlayer';
+import {Player} from './structures/UNO/Player';
 import {RichEmbed, Attachment} from 'discord.js';
 import Canvas from 'canvas';
-import fs from 'fs';
 
 const s = n => n == 1 ? '' : 's';
 
@@ -156,161 +156,5 @@ export class UNO extends Game {
         for (const card of this.deck.filter(c => c.wild))
             card.color = undefined;
         this.send('O baralho foi embaralhado.');
-    }
-}
-
-class Player extends AbstractPlayer {
-    constructor(user, game) {
-        super(user, game);
-        
-        this.hand = [];
-        this.called = false;
-        this.finished = false;
-    }
-
-    sortHand() {
-        this.hand.sort((a, b) => {
-            return a.value > b.value;
-        });
-    }
-
-    parseColor(color) {
-        switch ((color || '').toLowerCase()) {
-            case 'vermelho':
-            case 'red':
-            case 'r':
-                color = 'R';
-                break;
-            case 'amarelo':
-            case 'yellow':
-            case 'y':
-                color = 'Y';
-                break;
-            case 'verde':
-            case 'green':
-            case 'g':
-                color = 'G';
-                break;
-            case 'azul':
-            case 'blue':
-            case 'b':
-                color = 'B';
-                break;
-            default:
-                color = '';
-                break;
-        }
-        return color;
-    }
-
-    getCard(words) {
-        let color, id;
-        if (words.length === 1) {
-            id = words[0];
-        } else {
-            color = words[0];
-            id = words[1];
-        }
-        let _color = this.parseColor(color);
-        if (!_color) {
-            let temp = color;
-            color = id;
-            id = temp;
-            _color = this.parseColor(color);
-            if (!_color) {
-                this.game.send('Você precisa especificar uma cor! As cores são **red**, **yellow**, **green**, e **blue**.\n`ty.uno play <cor> <valor>`');
-                return null;
-            }
-        }
-        color = _color;
-        if (['WILD', 'WILD+4'].includes(id.toUpperCase())) {
-            let card = this.hand.find(c => c.id === id.toUpperCase());
-            if (!card) return undefined;
-            card.color = color;
-            return card;
-        } else {
-            return this.hand.find(c => c.id === id.toUpperCase() && c.color === color);
-        }
-    }
-
-    send(content) {
-        this.user.send(content);
-    }
-
-    async sendHand(turn = false) {
-        this.sortHand();
-        let len = (this.hand.length * 112) + 130,
-            canvas = new Canvas(len, 362),
-            ctx = canvas.getContext('2d');
-                
-        for (let i in this.hand) {
-            let card = this.hand[i],
-                image = new Canvas.Image;
-            image.src = card.URL;
-            ctx.drawImage(image, i * 112, 0, 242, 362);
-        }
-        await this.send(new RichEmbed()
-            .setAuthor('UNO', 'https://i.imgur.com/Zzs9X74.png')
-            .setDescription(`${turn ? "É seu turno! " : ''}Aqui está sua mão:\n\n${this.hand.map(h => `**${h}**`).join(' | ')}\n\nVocê tem ${this.hand.length} carta${s(this.hand.length)}.`)
-            .setThumbnail('attachment://cards.png')
-            .setColor('#E67E22')
-            .attachFile(new Attachment(canvas.toBuffer(), 'cards.png')));
-    }
-}
-
-class Card {
-    constructor(id, color) {
-        this.id = id;
-        this.wild = false;
-        this.color = color;
-        if (!this.color) this.wild = true;
-    }
-
-    get colorName() {
-        return {
-            R: 'Vermelho',
-            Y: 'Amarelo',
-            G: 'Verde',
-            B: 'Azul'
-        }[this.color];
-    }
-
-    get colorCode() {
-        return {
-            R: '#ff5555',
-            Y: '#ffaa00',
-            G: '#55aa55',
-            B: '#5555ff'
-        }[this.color] || '#080808';
-    }
-
-    get URL() {
-        return fs.readFileSync(`src/img/UNO/${this.color || ''}${this.id}.png`);
-    }
-
-    get value() {
-        let val = 0;
-        switch (this.color) {
-            case 'R': val += 100000; break;
-            case 'Y': val += 10000; break;
-            case 'G': val += 1000; break;
-            case 'B': val += 100; break;
-            default: val += 1000000; break;
-        }
-        switch (this.id) {
-            case 'SKIP': val += 10; break;
-            case 'REVERSE': val += 11; break;
-            case '+2': val += 12; break;
-            case 'WILD': val += 13; break;
-            case 'WILD+4': val += 14; break;
-            default: val += parseInt(this.id); break;
-        }
-        return val;
-    }
-
-    toString() {
-        if (this.color)
-            return this.colorName + ' ' + this.id;
-        else return this.id;
     }
 }
